refactor: Improve intuitiveness of creation file handling
